+++ applications/test/List3/Test-List3.C
@@ -47,15 +47,15 @@ using namespace Foam;
 template<class T>
 void printAddress(const UList<T>& list)
 {
-    Info<< "list addr: " << long(&list)
-        << " data addr: " << long(list.cdata()) << nl;
+    Info<< "list addr: " << uintptr_t(&list)
+        << " data addr: " << uintptr_t(list.cdata()) << nl;
 }
 
 
 template<class T>
 void printAddress(const SLList<T>& list)
 {
-    Info<< "list addr: " << long(&list)
+    Info<< "list addr: " << uintptr_t(&list)
         << " data addr: ???" << nl;
 }
